fastAPI/requirements.txt
fastapi
pydantic
requests
spotipy
google-cloud-pubsub
uvicorn